Organize all address logic into a single directory
